+++ libraries/microsoft-agents-activity/microsoft_agents/activity/entity/entity_types.py
@@ -8,6 +8,7 @@ class EntityTypes(str, Enum):
     """Well-known enumeration of entity types."""
 
     ACTIVITY_TREATMENT = "activityTreatment"
+    AI_CITATION = "https://schema.org/Message"
     GEO_COORDINATES = "GeoCoordinates"
     MENTION = "mention"
     PLACE = "Place"